Tratamento de erros

+++ src/main/java/com/own/portfolio/service/PostService.java
@@ -4,12 +4,15 @@ import com.own.portfolio.model.Post;
 import com.own.portfolio.database.PostRepository;
 import com.own.portfolio.utils.Image;
 import com.own.portfolio.utils.Token;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.nio.file.Path;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Objects;
 
 @Service
 public class PostService{
@@ -27,17 +30,28 @@ public class PostService{
         postRepository.savePost(post);
     }
 
-    public void editImage(MultipartFile image, int id) throws IOException, SQLException {
-        Image.saveFile(image);
-        Post post = onePost(id);
-        final String token = Token.getToken(post.getTitle());
-        Image.saveAtBucket(token);
-        post.setImage("https://erik-na-portfolio.s3.amazonaws.com/"+token);
-        editPost(post, id);
+    public int editImage(MultipartFile image, int id) throws IOException, SQLException {
+        if(image == null) return 400;
+
+        Path path = Image.saveFile(image);
+        Post oldPost = onePost(id);
+
+        if(oldPost == null) return 406;
+
+        final String token = Token.getToken(oldPost.getTitle());
+
+        assert path != null;
+        Image.saveAtBucket(token, path);
+        oldPost.setImage("https://erik-na-portfolio.s3.amazonaws.com/"+token+".jpg");
+        editPost(oldPost, id);
+        return 200;
     }
 
-    public void editPost(Post post, int id) throws SQLException {
+    public int editPost(Post post, int id) throws SQLException {
         Post oldPost = postRepository.onePost(id);
+
+        if(oldPost == null) return 406;
+        if(Objects.equals(post.getTitle(), oldPost.getTitle()) && Objects.equals(post.getDescription(), oldPost.getDescription())) return 400;
 
         oldPost.setImage(post.getImage());
         oldPost.setTitle(post.getTitle());
@@ -45,9 +59,12 @@ public class PostService{
         oldPost.setCreatedAt(post.getCreatedAt());
 
         postRepository.editPost(oldPost, id);
+        return 200;
     }
 
-    public void deletePost(int id) throws SQLException {
+    public int deletePost(int id) throws SQLException {
+        if(postRepository.onePost(id) == null) return 406;
         postRepository.deletePost(id);
+        return 200;
     }
 }
